Fix display name, add vanish, tweak flare
In this commit, I found a method that parses the the player name in a more natural form (thanks @cindyker). So we are now formatting the player's name correctly.

Additionally, I added an untested form of the vanish ability that gives the player invisibility for about two minutes, but takes away a notable chunk of their health.

The flare and vanish abilities also now check to see if the player's health is extremely low. The command will only kill the player if their health is equal to or below 1.0. If it isn't, the player will just be close to dead.

I also fixed a small issue with typing the command via console. Check your order first, kids!