Add BarAPI messages, fix flare power, clean up start listener
Officially added BarAPI support! If it's loaded on the server, it should use the bar for broadcasts.

Flare power was fixed. It was throwing the stacktrace because the command sender was considered a nearby player. Added some extra conditions to fix.

Start listener was cleaned up a good bit, but not production ready. It's next up on my agenda.